Add a required header

+++ ios.c
@@ -1,11 +1,11 @@
+#include <stdint.h>
 #include <stdio.h>
-#include <math.h>
 
 void smart() {
 	printf("Hello world!");
 }
 
-uint8_t b32decode(char* s) {
-	return (s)? *s : s;
+uintptr_t b32decode(char* s) {
+	return (s)? *s : (uintptr_t) s;
 	// TODO: small chance of overflowing on systems with more than 256 cells of memory
 }
